refactor: components 디렉토리 내부 파일 v1 > base로 이동
